Moved SearchAsyncTask to it's own file
